pyproject.toml
[project]
name = "agentic-fraud-investigator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "reportlab>=4.4.9",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]
